FIxed email confirmation logic (#38)
* FIxed email confirmation logic

* additional fix
